Update modules lesson after presentation
