+++ 03_modules/02.rb
@@ -8,6 +8,7 @@ class Animal
   end
 end
 
+# class can only inherit from one class
 class Cat < Animal
   def meow
     puts "#{name} makes meow!"
@@ -25,10 +26,11 @@ end
 # and meow method from Cat class
 lion = Lion.new('Mustafa')
 
+# both methods will work
 lion.meow
 lion.growl
 
-puts Lion.superclass
-puts Lion.superclass.superclass
-puts Lion.superclass.superclass.superclass
-puts Lion.superclass.superclass.superclass.superclass
+puts "Lion.superclass is: #{Lion.superclass} and has #{Lion.superclass.instance_methods.count} methods"
+puts "Cat.superclass is: #{Cat.superclass} and has #{Cat.superclass.instance_methods.count} methods"
+puts "Animal.superclass is: #{Animal.superclass} and has #{Animal.superclass.instance_methods.count} methods"
+puts "Object.superclass is: #{Object.superclass} and has #{Object.superclass.instance_methods.count} methods"
